requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyMuPDF==1.22.5
Pillow==10.0.1
gunicorn==21.2.0
psutil==5.9.5
flask-cors==4.0.0
memory-profiler==0.61.0  # For memory profiling
numpy==1.24.3  # For better array handling
python-dotenv==1.0.0  # For environment variables